requirements.txt
requests==2.31.0
types-requests==2.31.0.20240311
python-dotenv==1.0.0
click==8.1.7
assemblyai==0.41.0b4
websocket-client==1.7.0
pyaudio==0.2.14